Fix int32 overflow, thread safety, SQL injection, and harden exception handling across codebase
Fix order ID int32 overflow, bincount int8 overflow, CDF boundary bug, web route thread safety (deepcopy under lock), SQL Unicode paths and OBJECT_ID escaping, TMDL path injection, bare except→specific types, ConfigError over RuntimeError, and tighten test assertions.
Update CLAUDE.md with gotchas for int32 overflow, numpy bincount, CDF boundary, SQL escaping, and TMDL injection

+++ src/facts/sales/sales_logic/core/delivery.py
@@ -209,7 +209,7 @@ def compute_dates(rng, n, product_keys, order_ids_int, order_dates):
     if has_orders:
         # Any delayed line → order delayed
         delayed_any = (
-            np.bincount(inv_idx, weights=delayed_line.astype(np.int8), minlength=len(unique_orders)) > 0
+            np.bincount(inv_idx, weights=delayed_line.astype(np.float64), minlength=len(unique_orders)) > 0
         )
         is_order_delayed = delayed_any[inv_idx].astype(np.int8, copy=False)
     else:
